CY-2541 - Jenkins build fix  (#984)
* Removed brotli from size-limit. Improved dist folder deletion in production build.
* Removed RPM uploading to S3

+++ webpack.config.js
@@ -120,7 +120,7 @@ module.exports = (env, argv) => {
                   })
               ];
 
-    if (isProduction) {
+    if (isProduction && fs.existsSync(outputPath)) {
         try {
             fs.rmdirSync(outputPath, { recursive: true });
         } catch (err) {
